refactor: move header to SharedModule
