--- src/app/core/pages/header/header.component.html
@@ -1,1 +1,0 @@
-<div class="header-wrapper"></div>
